+++ src/components/ChatPanel.vue
@@ -2,7 +2,7 @@
 import { ScrollAreaRoot, ScrollAreaScrollbar, ScrollAreaThumb, ScrollAreaViewport } from 'reka-ui'
 import { computed, markRaw, nextTick, ref, watch } from 'vue'
 
-import APIKeySetup from '@/components/chat/APIKeySetup.vue'
+import ProviderSetup from '@/components/chat/ProviderSetup.vue'
 import ChatInput from '@/components/chat/ChatInput.vue'
 import ChatMessage from '@/components/chat/ChatMessage.vue'
 import { useAIChat } from '@/composables/use-chat'
@@ -41,7 +41,7 @@ function handleStop() {
 
 <template>
   <div data-test-id="chat-panel" class="flex min-w-0 flex-1 flex-col overflow-hidden">
-    <APIKeySetup v-if="!isConfigured" />
+    <ProviderSetup v-if="!isConfigured" />
 
     <template v-else>
       <ScrollAreaRoot class="min-h-0 flex-1">